requirements.txt
anyio==4.6.2.post1
asgiref==3.8.1
certifi==2024.8.30
Django==5.1
django-appconf==1.0.6
h11==0.14.0
httpcore==1.0.7
httpx==0.27.2
idna==3.10
pillow==10.4.0
python-telegram-bot==21.7
sniffio==1.3.1
sqlparse==0.5.1
tzdata==2024.1
whitenoise==6.7.0
